Adds swift markdown parsing project
We can use this rather than creating an NSTask that calls out to the
python library. This not only makes it safer (i.e. we know the
library must exist because we include it as a source file), it also
makes it a lot more concise... none of this 'waitForExecution'
malarkey on NSTask and then using NSPipes to capture the output.